Add -Knofz compile option for half-precision in shallow water model
This turns off automatic flush-to-zero behaviour for floating-point calculations.
Without this option, flush-to-zero is turned on automatically by -Kfast, so
numbers that occur outside of the normal half-precision range are set to zero. The
shallow water model was not able to advance when this was the case as such numbers
are common.